attributes/default.rb
I added a new attribute to be used by the recipe proxy.rb, this attribute defines which network interface liberty should be reached on, it defaults to eth0.
I also moved the two attributes only used by proxy.rb into their own namespace.

proxy.rb

I corrected the way that it looked for ipaddresses on chef nodes, and I changed the name of the apache config file so that apache will load it.

I created a new recepy: proxyText.rb which configres a Liberty server that serves a single jsp page - showip.jsp, the jsp page says which ip address it was served from. This was based on a design by Ant, so thank you ant.